fix(vue): restore raw input resolution
The v2 core rewrite removed client-side Vue input resolution while preserving reactive VueHeadClient input types. Resolve Vue values during rendering without adding watcher ownership to raw entries.
